continuous/requirements.txt
Pillow
scipy
tqdm
gradio
matplotlib
numpy
pandas
gradio
scikit-learn